Add support for derived peripherals
A derived peripheral is one whose peripheral declaration contains
a "derivedFrom" attr.  To support this, we cache all peripheral objects
in a table so a derived peripheral can copy its base.

A derived peripheral should then overwrite the fields copied fields
of the base, but that is future work.  For now, we just copy the name
and baseAddress since that is what is in the test file.