Improve setField
Added more to the doc comment.
Removed an assert that fired when a fieldVal
did not not fit within the given bitWidth.
This was removed because it could not be determined at compile time.
Added a TODO to figure out how to handle this during runtime.
For now, the bit mask will truncate any bits outside of the mask.