Update integer parsing
Some integers needed to have their type specified from int to uint32 because the tool compiles for the desktop where ints are 64-bit, but we want these ints to work on ARM Cortex CPUs, which are 32-bit.
Also added the ability to parse binary integers.  This is incomplete as it does not yet support don't-care bits indicated by 'x' in the number.