Add support for enums
Added support for enums in the parser, renderer and templates.
The SVD file declares an enum to represent possible values
a field in a register may take.  When an enum is declared,
a Nim enum is output and a new proc is generated to set
the register's field to the enum's value.
Updated