Add support for derived registers
During parsing, a baseRegister is the example for derived registers to follow.
In the parser, create a registerCache and pass it down to parseSvdRegister().
A derived register keeps a reference to the base register.

When rendering a register, we use one of two rendering procedures.
If the register has a reference to a base register, it is a derived register,
and we call renderDerivedRegister().  Otherwise, we call renderDistinctRegister().

In the metagenerator, the two declare*Register() templates generate mostly the same code.
The only difference is the parent of a derived register object is the base register object.
This allows all registers derived from a common base to be type-compatible.

Added tests for parsing a derived register.