Rewrite the parser to use a declarative spec
While writing the previous version of this tool, I realized that each
parsing function was doing very similar work, just with different field
names and types.  So I moved the field names and types to const
declarations in the svd_spec file and wrote a generalized parser func.

The svd_spec module declares the structures of the SVD specification.
The parser uses a recursive function that utilizes the spec to generate
a tree of SvdElementValue representing the same information as
the input .svd file.

During parsing, no value parsing is done on the data fields;
the intermediate values remain in string form.  This reduces
processing.