Copy metagenerator to target package
It was becoming difficult to use the tool because of the dependency
on metagenerator.nim.  The solution was to copy it to the same
folder holding all of the generated <peripheral>.nim modules.

Also fixed the unit tests.  Mostly to fix having a space in the path
of an executable command.

Fixes #5