Fix boolean device fields that were rendered 0/1
Created a compile-time func to inspect the spec and
determine when a field type is boolean.  Then use that
in a template to cast a boolean to false/true.

Note that the input from the SVD can be 0/1 or false/true.

This is only used in the device module's FPU_ and MPU_PRESENT
at this time.