Clean up tests
The primary purpose was to stop the cross-building between
project, tests and examples.  Let them each build for themselves.

The secondary purpose was to be sure to auto-clean
packages and files built during tests and examples.

Removed a test file that did nothing.
Removed an .svd file that was not used.
Removed an after build nimble task that built a target for an example;
the example should have its own build task (done in a parallel branch).
Renamed the STM .svd file to avoid name collisions in the project.
Consolidated big .svd render tests to one, test.svd.
Made the smaller .svd test even smaller for more timely tests.